+++ opencood/tools/inference_utils.py
@@ -68,7 +68,7 @@ def inference_early_fusion(batch_data, model, dataset):
 
     processed_lidar = cav_content['processed_lidar']
     voxel_features = processed_lidar['voxel_features'].to(torch.float32)
-    voxel_coords = processed_lidar['voxel_coords'].to(torch.int32)
+    voxel_coords = processed_lidar['voxel_coords'].to(torch.float32)
     voxel_num_points = processed_lidar['voxel_num_points'].to(torch.int32)
 
     psm, rm = model(voxel_features, voxel_coords, voxel_num_points, record_len,
