+++ opencood/tools/train.py
@@ -153,7 +153,7 @@ def main():
 
             processed_lidar = cav_content['processed_lidar']
             voxel_features = processed_lidar['voxel_features'].to(torch.float32)
-            voxel_coords = processed_lidar['voxel_coords'].to(torch.int32)
+            voxel_coords = processed_lidar['voxel_coords'].to(torch.float32)
             voxel_num_points = processed_lidar['voxel_num_points'].to(torch.int32)
 
             # case1 : late fusion train --> only ego needed,
@@ -212,7 +212,7 @@ def main():
 
                     processed_lidar = cav_content['processed_lidar']
                     voxel_features = processed_lidar['voxel_features'].to(torch.float32)
-                    voxel_coords = processed_lidar['voxel_coords'].to(torch.int32)
+                    voxel_coords = processed_lidar['voxel_coords'].to(torch.float32)
                     voxel_num_points = processed_lidar['voxel_num_points'].to(torch.int32)
 
                     output_dict = model(voxel_features, voxel_coords, voxel_num_points, record_len,
